+++ src/db/migration/3_create_table_dishIngredient.sql
@@ -2,7 +2,7 @@ CREATE TABLE if not exists dish_ingredient
     (
         id_dish INT PRIMARY KEY,
         id_ingredient INT PRIMARY KEY,
-        required_quality INT,
+        required_quality NUMERIC,
         unit unit,
         FOREIGN (id_dish) REFERENCES dish(id),
         FOREIGN (id_ingredient) REFERENCES ingredient(id)
